implement bobocard in list and bobopage
